requirement.txt
pytorch==1.11
cudatoolkit==11.3
tqdm==4.64
normflows==1.4
dgl==0.9.0
tensorboardx==2.5.1